serverML/api/serializers.py
from .models import Prompt, Branch, Block
from rest_framework import serializers

class PromptSerializer(serializers.ModelSerializer):
    class Meta:
        model = Prompt
        fields = ('id', 'prompt_title', 'text', 'next')

class BranchSerializer(serializers.ModelSerializer):
    class Meta:
        model = Branch
        fields = ('id', 'prompt')

class BlockSerializer(serializers.ModelSerializer):
    class Meta:
        model = Block
        fields = ('id', 'branch', 'text', 'next')
