serverML/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2==2.9.9
transformers==4.40.2